updated styling on input, slight modification to tests
